judge_supervise and balloon is now possible to give permissions to, and available for judges
